chore: include React and Styled-Components
React is a JavaScript library for building user interfaces. It uses a
declarative paradigm and aims to be both efficient and flexible.
styled-components is a JavaScript library for styling React
applications. It removes the mapping between styles and components, and
lets you write actual CSS augmented with JavaScript.

Styled-Components is a JavaScript library for styling React
applications. It removes the mapping between styles and components, and
lets you write actual CSS augmented with JavaScript.